backend/requirements.txt
# ── Web Framework ─────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9

# ── Environment ───────────────────────────────────────────────────────────────
python-dotenv>=1.0.0

# ── LLM ───────────────────────────────────────────────────────────────────────
groq>=0.5.0

# ── PDF ───────────────────────────────────────────────────────────────────────
pdfplumber>=0.10.3
pypdf>=4.1.0

# ── RAG ───────────────────────────────────────────────────────────────────────
langchain>=0.1.0
langchain-community>=0.0.20
langchain-huggingface>=0.0.3
langchain-text-splitters>=0.0.1

# ── Embeddings ────────────────────────────────────────────────────────────────
sentence-transformers>=2.6.1
transformers>=4.38.0

# ── Vector Store ──────────────────────────────────────────────────────────────
faiss-cpu>=1.8.0

# ── Vision ────────────────────────────────────────────────────────────────────
torch>=2.2.0
Pillow>=10.2.0

# ── Export ────────────────────────────────────────────────────────────────────
reportlab>=4.1.0

# ── Utilities ─────────────────────────────────────────────────────────────────
numpy>=1.26.0